Add DropDown widget
Resolves #79
